Create  Dijkstra's algorithm.cpp
Dijkstra's algorithm is used to find the shortest path from a starting node to all other nodes in a weighted graph. Here's a step-by-step explanation:

Step 1: Initialization

Choose a starting node. This will be the source node.
Assign a tentative distance value to every node in the graph. Initialize the distance of the source node to 0 and the distance of all other nodes to infinity.
Maintain a set of unvisited nodes, initially containing all nodes.
Step 2: Selecting the Closest Node

From the set of unvisited nodes, select the node with the smallest tentative distance. This node will be the current node.
Step 3: Update Distances

For the current node, consider all of its unvisited neighbors.
Calculate their tentative distances by adding the weight of the edge from the current node to each neighbor to the current node's tentative distance.
If this new tentative distance is less than the current distance of the neighbor, update the neighbor's distance to this new value.
Step 4: Mark as Visited

Once all neighbors of the current node have been considered, mark the current node as visited (remove it from the set of unvisited nodes).
Step 5: Repeat

Repeat steps 2-4 until either all nodes have been visited or the smallest tentative distance among the unvisited nodes is infinity (indicating that there is no path between the source node and that node).
Step 6: Result

Once all nodes have been visited or no more progress can be made, the algorithm terminates. At this point, the tentative distances calculated for each node represent the shortest path distance from the source node.